requirements.txt
Flask>=3.0.0
python-gnupg>=0.5.0
Flask-WTF>=1.2.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
